+++ src/components/Home/FeaturedPapers.js
@@ -118,6 +118,7 @@ const FeaturedPapers = ({ researchFieldId }) => {
                                             ...paper,
                                             title: paper.label
                                         }}
+                                        showAddToComparison={false}
                                         key={`pc${paper.id}`}
                                     />
                                 );