+++ src/components/SmartReview/ViewArticle.js
@@ -4,8 +4,9 @@ import Tippy from '@tippyjs/react';
 import { toggleHistoryModal as toggleHistoryModalAction } from 'actions/smartReview';
 import Acknowledgements from 'components/SmartReview/Acknowledgements';
 import AuthorsList from 'components/SmartReview/AuthorsList';
-import MarkdownRenderer from 'components/SmartReview/MarkdownRenderer';
 import SectionDataTable from 'components/SmartReview/DataTable/SectionOntology';
+import MarkdownRenderer from 'components/SmartReview/MarkdownRenderer';
+import Outline from 'components/SmartReview/Outline';
 import ListReferences from 'components/SmartReview/References/ListReferences';
 import SectionVisualization from 'components/SmartReview/SectionVisualization';
 import { SectionStyled } from 'components/SmartReview/styled';
@@ -31,12 +32,12 @@ const ViewArticle = () => {
     const dispatch = useDispatch();
     const latestVersionId = versions?.[0]?.id;
     const newVersionAvailable = isPublished && latestVersionId !== id;
-
     const toggleHistoryModal = () => dispatch(toggleHistoryModalAction());
 
     return (
         <>
-            <Container className="print-only p-0">
+            <Container className="print-only p-0" style={{ position: 'relative' }}>
+                <Outline />
                 {!isPublished && (
                     <Alert color="warning" fade={false} className="box">
                         Warning: you are viewing an unpublished version of this article. The content can be changed by anyone.{' '}
@@ -81,7 +82,12 @@ const ViewArticle = () => {
                                 ) {
                                     return (
                                         <section key={section.id} typeof="doco:Section">
-                                            <h2 className="h4 border-bottom mt-5" typeof="doco:SectionTitle" property="c4o:hasContent">
+                                            <h2
+                                                id={`section-${section.id}`}
+                                                className="h4 border-bottom mt-5"
+                                                typeof="doco:SectionTitle"
+                                                property="c4o:hasContent"
+                                            >
                                                 {section.title.label}
                                             </h2>
                                             {section.type.id === CLASSES.ONTOLOGY_SECTION && <SectionDataTable key={section.id} section={section} />}
@@ -142,6 +148,7 @@ const ViewArticle = () => {
                                                 style={{ whiteSpace: 'pre-line' }}
                                                 typeof="doco:SectionTitle"
                                                 property="c4o:hasContent"
+                                                id={`section-${section.id}`}
                                             >
                                                 {section.title.label}
                                             </h2>
@@ -151,7 +158,12 @@ const ViewArticle = () => {
                                 }
                             })}
                             <section typeof="doco:Section deo:Acknowledgements">
-                                <h2 className="h4 border-bottom mt-5" typeof="doco:SectionTitle" property="c4o:hasContent">
+                                <h2
+                                    id="section-acknowledgements"
+                                    className="h4 border-bottom mt-5"
+                                    typeof="doco:SectionTitle"
+                                    property="c4o:hasContent"
+                                >
                                     <Tippy content="Acknowledgements are automatically generated based on ORKG users that contributed to resources used in this article">
                                         <span>Acknowledgements</span>
                                     </Tippy>
@@ -160,7 +172,7 @@ const ViewArticle = () => {
                             </section>
 
                             <section typeof="doco:Section deo:Reference">
-                                <h2 className="h4 border-bottom mt-4" typeof="doco:SectionTitle" property="c4o:hasContent">
+                                <h2 id="section-references" className="h4 border-bottom mt-4" typeof="doco:SectionTitle" property="c4o:hasContent">
                                     References
                                 </h2>
                                 <ListReferences />