+++ src/pages/ResearchFields/ResearchFields.js
@@ -61,6 +61,25 @@ const ResearchFields = () => {
             </Container>
             <Container className="p-0">
                 <div className="box rounded-lg p-4">
+                    <div className="d-flex">
+                        <div>
+                            <p>
+                                The <i>ORKG Research fields taxonomy</i> is used to organize and facilitate browsing and exploring the research
+                                knowledge graph.
+                            </p>
+                            <div>
+                                Further information about the taxonomy can be found in the{' '}
+                                <a
+                                    href="https://gitlab.com/TIBHannover/orkg/orkg-frontend/-/wikis/ORKG-Research-fields-taxonomy"
+                                    target="_blank"
+                                    rel="noopener noreferrer"
+                                >
+                                    ORKG wiki
+                                </a>
+                            </div>
+                        </div>
+                    </div>
+                    <hr />
                     <Row>
                         <Col md="5" className="border-right">
                             <ResearchFieldSelector